refactor: move useStagingDragInteraction to interactions layer
Move useStagingDragInteraction hook from feature-specific location
(src/features/staging/hooks/) to the app-level interactions layer
(src/hooks/interactions/) where it belongs with other interaction
mode handlers (draw, drag, resize).

This fixes a cross-feature import violation where grid-editor was
importing from the staging feature. The interaction system is
app-level infrastructure, and all mode handlers should be colocated.

Changes:
- Move useStagingDragInteraction.ts to src/hooks/interactions/
- Update barrel export to use local import instead of re-export
- Remove empty staging/hooks directory
- Update consuming imports in useInteraction.ts
