--- src/features/staging/hooks/index.ts
@@ -1,2 +1,0 @@
-// Barrel export for staging hooks
-export { useStagingDragInteraction } from './useStagingDragInteraction';
