+++ src/features/staging/index.ts
@@ -1,3 +1,2 @@
 // Barrel export for staging feature
 export * from './components';
-export * from './hooks';
